bug: prevent infinite loop in credit note form (#2689)
## Context

A recent update changed this useEffect watch list. 
This useEffect now watch a form and at the same time update its value,
generating an infinite loop.

## Description

This pull request makes sure we do not watch the form object entirely.
It is still working because all the necessary values are watched, and
the value will be correct when the effect will be fired.
I added a note to make sure we don't make the same mistake later. We'll
need to refactor this component to make sure the design pattern of it
does not allow search issues.

<!-- Linear link -->
Fixes ISSUE-1229