feat(integrations): allow integration mapping per billing entity (#2668)
## Context

Integration mappings allow the use of third party services. But
currently, we are limiting this mapping at one per connection.
Now, we will allow the same connection to a third party to have multiple
mappings depending on Lago's billing entity

## Description

Updates the mapping drawers to allow defining multiple mappings
depending on the billing entity.
the previous mappings will be defined as "Default mapping". If a billing
entity doesn't have a mapping, it will fallback to the default one

Backend can be found
[here](https://github.com/getlago/lago-api/pull/4458)


<!-- Linear link -->
Fixes
[ISSUE-1248](https://linear.app/getlago/issue/ISSUE-1248/allow-integration-mapping-per-billing-entity)