misc: better query and state management on organization switch (#2704)
## Context

The organization switch
[seems](https://lago-eu.sentry.io/issues/77122518/?alert_rule_id=172021&alert_type=issue&notification_uuid=a72a1998-485a-47d5-ab69-b7212016aeea&project=4509111387619408)
to trigger unhandled error
This is an attempt to handle it in a better way.

## Description

This PR extracts the orga switch logic, wrap with a try catch and
manually send to sentry error cause the error object there does not
bubbles correctly.

Been aggressive in the disabled state using a `useState` boolean to
prevent any double click that could occur before the promise runs.

Also improved some ternary in template, removed an unused error and
improved cache/query state used in this file.