bug: improve activity logs unavailable error state (#2778)
## Context

User can access an activity logs list within the application.

This list is available on different objects (user, plan, billable
metrics, ...) but it also exist a "general" activity logs tab, where you
can apply filters to refine results.

This feature is only available for premium user AND those who have
performed the setup with Lago team to have this feature working.
So it can happen our users have a paying license but does not have the
exact configuration yet to have access to this feature.

In such case, we currently display the default error state.

## Description

This PR make sure we catch the correct error and display a better
placeholder message.

<!-- Linear link -->
Fixes ISSUE-1258