bug(Invoice): fetch correct payload on refresh (#2812)
## Context

When an invoice is in draft mode, you can keep sending events for it,
and you can trigger a refresh action that will fetch again those new
fees.

Behind the hood, the backend is destroying and recreating those fees
with the units being protected.

Today, this endpoint does not fetch the correct payload to update all
the rows

## Description

This pull request simply updates the payload requested by this endpoint.
Also note that the `AllInvoiceDetailsForCustomerInvoiceDetails` fragment
is still used, so that's why you don't see it removed to this pull
request.

<!-- Linear link -->
Fixes ISSUE-1311