chore: rework caching logic (#2801)
## Context

The cache policy is very verbose and kinda hidden within the
configuration of Apollo.
Every time we have a new collection, we need to define the cache key to
make sure the objects are merged within the same list to be displayed in
the infinite scroll list.
It also requires us to ensure that we are watching the correct arguments
from the object to help Apollo in proving the cache performance.
This can be very tricky/ not very handy/ forgotten during any
implementation or improvement of a feature.
A good example is when you have the addition of a filter within a list.
You may forget to add this new attribute name within the cache list to
be better observed by Apollo and set within a specific cache location.

## Description

This pull request makes sure that we are automatically watching all the
attributes that are passed within the query payload input.
It helps preventing us to manually update this list and makes Apollo
kind of better at caching the data results.

Another side effect is that we can slowly start to remove the custom
fetch policy and next fetch policies within those list pages.
Those were initially placed here because it was kind of flaky when you
were switching from a cool filter to another, and it was better to use
`false` as a key, but down the line we were relying on Apollo logic to
catch if a query result was to be cached or not. Which I doubt it was
good at.

It also makes the definition less verbose and includes the pagination
logic.

I tried to keep some of the queries "special treatment" here in order to
not break any implementation, those should at some pont be better
managed for sure.