bug: prevent showing skip rule button if no rule is defined (#2851)
## Context

When setting top-up min or max limits, you can choose to bypass the
rules.

In the UI, it shows a ON/OFF switch to enable or not this rule.

However it's visible in some cases that does not makes sense

## Description

This PR ensure the bypass switch is only visible if a min or max rule
have been defined.

It's currently flaky as it does not check for all possible values, such
as empty string (when you enter a min/max value then reset the input)

<!-- Linear link -->
Fixes ISSUE-1312