misc(CreditNotes) simplify data shape used in form (#2874)
## Context

We are planning to add some functionalities to the Credit Notes form. 
Today the form is shaping the items in the form to be used in a way that
is no longer really changing anything visually.
This pull request is trying to simplify this while maintaining the same
functionalities.

## Description

This contains two commits:
1. Add the tests to make sure we're not breaking anything.
2. Changes the way we group and format the items before being displayed.
Thus, we just changed the data shape in the test, and they still work
after that.